tests(resolver): Group integration test with other minimal version test
